Fixed lib work; Fixed gradle; Fixed deploy; Changed package to fun.kotlingang.kds; Deployed; New more convenient API introduced;
