+++ buildSrc/src/main/kotlin/AppInfo.kt
@@ -1,6 +1,6 @@
 object AppInfo {
     const val PACKAGE = "com.kotlingang.kds"
-    const val VERSION = "1.2.11"
+    const val VERSION = "2.5"
     const val ARTIFACT_ID = "kds"
     const val NAME = "Kotlin Data Storage"
     const val DESCRIPTION = "Multiplatform Coroutine-Based Kotlin Library for storing data via kotlinx.serialization"
