+++ src/browserMain/kotlin/fun/kotlingang/kds/storage/StorageUtils.kt
@@ -1,0 +1,6 @@
+package `fun`.kotlingang.kds.storage
+
+// Nothing because Browser js using localStorage for saving data
+actual val dirPath = ""
+
+actual fun String.joinPath(path: String) = "$this/$path"
