--- src/browserMain/kotlin/package.kt
@@ -1,1 +1,0 @@
-package com.kotlingang.kds
