+++ src/commonMain/kotlin/fun/kotlingang/kds/storage/BaseStorage.kt
@@ -1,0 +1,9 @@
+package `fun`.kotlingang.kds.storage
+
+
+internal expect class BaseStorage(path: String) {
+    val path: String
+
+    suspend fun saveStorage(text: String)
+    suspend fun loadStorage(): String?
+}
