+++ src/commonMain/kotlin/fun/kotlingang/kds/storage/StorageUtils.kt
@@ -1,0 +1,5 @@
+package `fun`.kotlingang.kds.storage
+
+
+expect val dirPath: String
+expect fun String.joinPath(path: String): String
