--- src/commonMain/kotlin/storage/BaseStorage.kt
@@ -1,9 +1,0 @@
-package com.kotlingang.kds.storage
-
-
-internal expect class BaseStorage(path: String) {
-    val path: String
-
-    suspend fun saveStorage(text: String)
-    suspend fun loadStorage(): String?
-}
