--- src/commonMain/kotlin/storage/StorageUtils.kt
@@ -1,5 +1,0 @@
-package com.kotlingang.kds.storage
-
-
-expect val dirPath: String
-expect fun String.joinPath(path: String): String
