+++ src/jvmMain/kotlin/fun/kotlingang/kds/storage/StorageUtils.kt
@@ -1,0 +1,8 @@
+package `fun`.kotlingang.kds.storage
+
+import java.nio.file.Paths
+
+
+actual val dirPath: String = System.getProperty("user.dir")
+
+actual fun String.joinPath(path: String) = Paths.get(this, path).toString()
