--- src/jvmMain/kotlin/storage/StorageUtils.kt
@@ -1,8 +1,0 @@
-package com.kotlingang.kds.storage
-
-import java.nio.file.Paths
-
-
-actual val dirPath: String = System.getProperty("user.dir")
-
-actual fun String.joinPath(path: String) = Paths.get(this, path).toString()
