--- utils/src/commonMain/kotlin/AnyUtils.kt
@@ -1,5 +1,0 @@
-package com.kotlingang.kds
-
-
-@Suppress("unused")
-val Any?.unit get() = Unit
