+++ utils/src/commonMain/kotlin/fun/kotlingang/kds/AnyUtils.kt
@@ -1,0 +1,5 @@
+package `fun`.kotlingang.kds
+
+
+@Suppress("unused")
+val Any?.unit get() = Unit
